requirements.txt
asgiref
Django
django-filter
djangorestframework
Markdown
pillow
sqlparse
tzdata